+++ cw-orch/tests/daemon_contract.rs
@@ -4,21 +4,30 @@ mod tests {
     /*
         DaemonAsync contract general tests
     */
-    use crate::common;
-    use cw_orch::prelude::*;
-    use std::sync::Arc;
+
+    use cw_orch::{environment::TxHandler, prelude::*};
+    use mock_contract::{InstantiateMsg, MigrateMsg, QueryMsg};
 
     use cosmwasm_std::Addr;
 
     use speculoos::prelude::*;
-    use tokio::runtime::Runtime;
 
     #[test]
     #[serial_test::serial]
     fn helper_traits() {
-        let runtime = Arc::new(Runtime::new().unwrap());
+        use cw_orch::prelude::networks;
 
-        let (sender, contract) = common::contract::start(&runtime);
+        let runtime = tokio::runtime::Runtime::new().unwrap();
+
+        let daemon = Daemon::builder()
+            .chain(networks::LOCAL_JUNO)
+            .handle(runtime.handle())
+            .build()
+            .unwrap();
+
+        let sender = daemon.sender();
+
+        let contract = mock_contract::MockContract::new("test:mock_contract", daemon.clone());
 
         asserting!("address is not present")
             .that(&contract.address())
@@ -32,7 +41,7 @@ mod tests {
             .that(&contract.latest_is_uploaded().unwrap())
             .is_true();
 
-        let init_msg = common::contract::get_init_msg(&sender);
+        let init_msg = &InstantiateMsg {};
 
         let _ = contract.instantiate(&init_msg, Some(&Addr::unchecked(sender)), Some(&[]));
 
@@ -43,7 +52,9 @@ mod tests {
         asserting!("migrate_if_needed is none")
             .that(
                 &contract
-                    .migrate_if_needed(&cw20_base::msg::MigrateMsg {})
+                    .migrate_if_needed(&MigrateMsg {
+                        t: "migrate".to_string(),
+                    })
                     .unwrap(),
             )
             .is_none();
@@ -61,7 +72,9 @@ mod tests {
         asserting!("migrate_if_needed is some")
             .that(
                 &contract
-                    .migrate_if_needed(&cw20_base::msg::MigrateMsg {})
+                    .migrate_if_needed(&MigrateMsg {
+                        t: "migrate".to_string(),
+                    })
                     .unwrap(),
             )
             .is_some();
@@ -74,9 +87,19 @@ mod tests {
     #[test]
     #[serial_test::serial]
     fn cw_orch_x() {
-        let runtime = Runtime::new().unwrap();
+        use cw_orch::prelude::networks;
 
-        let (sender, contract) = common::contract::start(&runtime);
+        let runtime = tokio::runtime::Runtime::new().unwrap();
+
+        let daemon = Daemon::builder()
+            .chain(networks::LOCAL_JUNO)
+            .handle(runtime.handle())
+            .build()
+            .unwrap();
+
+        let sender = daemon.sender();
+
+        let contract = mock_contract::MockContract::new("test:mock_contract", daemon.clone());
 
         // upload contract
         let upload_res = contract.upload();
@@ -88,25 +111,21 @@ mod tests {
 
         log::info!("Using code_id {}", code_id);
 
-        // init msg for contract
-        let init_msg = common::contract::get_init_msg(&sender);
-
         // instantiate contract on chain
-        let init_res = contract.instantiate(&init_msg, Some(&sender), None);
+        let init_res = contract.instantiate(&InstantiateMsg {}, Some(&sender), None);
         asserting!("instantiate is successful")
             .that(&init_res)
             .is_ok();
 
         // do a query and validate its successful
-        let query_res =
-            contract.query::<cw20::BalanceResponse>(&cw20_base::msg::QueryMsg::Balance {
-                address: sender.to_string(),
-            });
+        let query_res = contract.query::<String>(&QueryMsg::FirstQuery {});
         asserting!("query is successful").that(&query_res).is_ok();
 
         // validate migrations are successful
         let migrate_res = contract.migrate(
-            &cw20_base::msg::MigrateMsg {},
+            &MigrateMsg {
+                t: "migrate".to_string(),
+            },
             code_id.parse::<u64>().unwrap(),
         );
         asserting!("migrate is successful")
